Funcionalidade de adicionar administrador no JSON funcional
