+++ BancoDeDados/Administrador.json
@@ -1,1 +1,1 @@
-[{"ident":0,"login":"gabriel","senha":"gabriel4"}]
+[{"ident":0,"login":"gabriel","senha":"gabriel"},{"ident":0,"login":"rodrigo","senha":"rodrigo"},{"ident":0,"login":"predo","senha":"predo"},{"ident":0,"login":"cluba","senha":"cluba"},{"ident":5,"login":"cinco","senha":"cinco"}]
